requirements.txt
absl-py==2.3.1
addict==2.4.0
altair==5.5.0
attrs==25.4.0
basicsr @ git+https://github.com/xinntao/BasicSR.git@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
blinker==1.9.0
cachetools==6.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.1
colorama==0.4.6
filelock==3.19.1
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
idna==3.4
ImageIO==2.37.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
lmdb==1.7.5
Markdown==3.10
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==2.12.0
networkx==3.3
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
realesrgan @ git+https://github.com/xinntao/Real-ESRGAN.git@a4abfb2979a7bbff3f69f58f58ae324608821e27
referencing==0.37.0
requests==2.28.1
rpds-py==0.29.0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
sympy==1.14.0
tb-nightly==2.21.0a20251023
tenacity==9.1.2
tensorboard-data-server==0.7.2
tifffile==2025.5.10
toml==0.10.2
tomli==2.3.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.13
watchdog==6.0.0
Werkzeug==3.1.3
yapf==0.43.0
